+++ src/api/mod.rs
@@ -1,7 +1,10 @@
 //! API helpers to make use of k8s-openapi easier
 
+/// Empty struct for Void
+#[derive(Clone, Deserialize)]
+pub struct Discard {}
 /// Shortcut type for discarding one type parameter option
-pub type Void = Option<()>;
+pub type Void = Option<Discard>;
 
 mod reflector;
 pub use self::reflector::{
@@ -14,12 +17,15 @@ pub use self::informer::{
     Informer,
 };
 
+mod api;
+pub use api::{
+    Api,
+    GetParams,
+};
+
 mod resource;
 pub use self::resource::{
     Resource,
-    ApiResource,
-    ResourceType,
-    QueryParams,
     WatchEvent,
     ApiError,
 };